requirements.txt
fastapi==0.103.2
tortoise-orm==0.20.0
uvicorn==0.23.2
asyncpg==0.28.0
email_validator==2.0.0
slowapi==0.1.8
websockets==11.0.3
requests==2.29.0
celery==5.3.4
flower==2.0.1
redis==5.0.1
fastapi_pagination==0.12.10
uniswap_python==0.7.1
psycopg2_binary==2.9.9
aioredis==2.0.1
python_jose==3.3.0
gql==3.4.1
requests_toolbelt==1.0.0
boto3==1.28.80
python-multipart==0.0.6
tweepy==4.14.0
python-telegram-bot==20.7
cryptography==42.0.5